[update]修改了课程、实验的删除方法的逻辑

+++ src/main/java/com/gdou/teaching/controller/teacher/TeacherCourseController.java
@@ -293,9 +293,10 @@ public class TeacherCourseController {
     @GetMapping("/invalid/{courseId}")
     @Auth
     public ResultVO invalid(@PathVariable("courseId") Integer courseId) {
-        courseService.invalid(courseId);
-        //删除课程下的选课记录
-        achievementService.deleteAchievementByCourseId(courseId);
+        if(courseService.invalid(courseId)){
+            //删除课程下的选课记录
+            achievementService.deleteAchievementByCourseId(courseId);
+        }
         return ResultVOUtil.success();
     }
